Update cs205_final_exam.sh
Added in code for each of the three values. First one prints the number of lines in the file aka total number of pokemon. Second one takes the sum of column five and divides by the number of lines in the column. Third one does the same as the second one except in column six.